+++ test/core/mount.test.ts
@@ -14,6 +14,7 @@ describe('Mount', () => {
 				.then((x) => x.text())
 		).toBe('http://elysiajs.com/')
 	})
+
 	it('preserve request URL with query', async () => {
 		const plugin = new Elysia().get('/', ({ request }) => request.url)
 
@@ -24,5 +25,34 @@ describe('Mount', () => {
 				.handle(new Request('http://elysiajs.com/mount/?a=1'))
 				.then((x) => x.text())
 		).toBe('http://elysiajs.com/?a=1')
+	})
+
+	it('preserve body', async () => {
+		const handler = async (req: Request) => {
+			return new Response(await req.text())
+		}
+
+		const app = new Elysia()
+			.mount('/mount', (req) => handler(req))
+			.post('/not-mount', ({ body }) => body)
+
+		const options = {
+			method: 'POST',
+			headers: {
+				'content-type': 'text/plain'
+			},
+			body: 'sucrose'
+		}
+
+		const res = await Promise.all([
+			app
+				.handle(new Request('http://elysiajs.com/mount', options))
+				.then((x) => x.text()),
+			app
+				.handle(new Request('http://elysiajs.com/not-mount', options))
+				.then((x) => x.text())
+		])
+
+		expect(res).toEqual(['sucrose', 'sucrose'])
 	})
 })